requirements-test.txt
# for main app
osbot-fast-api-serverless
mgraph-ai-service-cache-client

# for integration tests
mgraph_ai_service_cache
mgraph_ai_service_html

# for pytest
pytest
pytest-cov

# for fastapi testing
httpx
requests